+++ httpsupgrade/httpsupgrade-impl/src/androidTest/java/com/duckduckgo/httpsupgrade/impl/BloomFilterTest.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2023 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.httpsupgrade.impl
+
+import androidx.test.platform.app.InstrumentationRegistry
+import com.duckduckgo.httpsupgrade.impl.BloomFilter.Config.ProbabilityConfig
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertTrue
+import org.junit.Test
+
+class BloomFilterTest {
+
+    private lateinit var testee: BloomFilter
+    private val context = InstrumentationRegistry.getInstrumentation().targetContext.applicationContext
+
+    @Test
+    fun whenBloomFilterEmptyThenContainsIsFalse() {
+        testee = BloomFilter(context, ProbabilityConfig(FILTER_ELEMENT_COUNT, TARGET_ERROR_RATE))
+        assertFalse(testee.contains("abc"))
+    }
+
+    @Test
+    fun whenBloomFilterContainsElementThenContainsIsTrue() {
+        testee = BloomFilter(context, ProbabilityConfig(FILTER_ELEMENT_COUNT, TARGET_ERROR_RATE))
+        testee.add("abc")
+        assertTrue(testee.contains("abc"))
+    }
+
+    companion object {
+        const val FILTER_ELEMENT_COUNT = 5000
+        const val TARGET_ERROR_RATE = 0.001
+    }
+}
